// dashboard-main/backup_temp/types.ts
export interface HeadToHeadConditionalFormatRule {
  id: string;
  criteria: 'specific_value' | 'column_dept_avg' | 'row_avg';
  operator: '>' | '<' | '=';
  value: number;
  textColor: string;
  backgroundColor: string;
}

export interface HeadToHeadTableConfig {
    id: string;
    tableName: string;
    selectedSubgroups: string[];
    selectedParentGroups?: string[];
    metricType: 'revenue' | 'quantity' | 'revenueQD' | 'hieuQuaQD';
    totalCalculationMethod?: 'sum' | 'average';
    conditionalFormats?: HeadToHeadConditionalFormatRule[];
}